+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeCopier.java
@@ -110,8 +110,8 @@ public class AnnotatedTypeCopier
 
     final AnnotatedDeclaredType copy = makeOrReturnCopy(original, originalToCopy);
 
-    if (original.wasRaw()) {
-      copy.setWasRaw();
+    if (original.isUnderlyingTypeRaw()) {
+      copy.setIsUnderlyingTypeRaw();
     }
 
     if (original.enclosingType != null) {