+++ checker/src/main/java/org/checkerframework/checker/mustcall/MustCallNoCreatesMustCallForChecker.java
@@ -5,7 +5,7 @@ import org.checkerframework.framework.source.SupportedOptions;
 
 /**
  * This copy of the Must Call Checker is identical, except that it does not load the stub files that
- * treat unconnected sockets as {@code @MustCall({})}. See SocketAccumulationFrames.astub.
+ * treat unconnected sockets as {@code @MustCall({})}. See SocketCreatesMustCallFor.astub.
  *
  * <p>The only difference is the contents of the @StubFiles annotation.
  */
@@ -13,5 +13,5 @@ import org.checkerframework.framework.source.SupportedOptions;
   "JavaEE.astub",
   "Reflection.astub",
 })
-@SupportedOptions({MustCallChecker.NO_ACCUMULATION_FRAMES})
-public class MustCallNoAccumulationFramesChecker extends MustCallChecker {}
+@SupportedOptions({MustCallChecker.NO_CREATES_MUSTCALLFOR})
+public class MustCallNoCreatesMustCallForChecker extends MustCallChecker {}