+++ docs/CHANGELOG.md
@@ -5,7 +5,7 @@ Version 3.14.0 (June 1, 2021)
 
 The Resource Leak Checker ensures that certain methods are called on an
 object before it is de-allocated. By default, it enforces that `close()` is
-called on any expression whose type might be `java.io.Closeable`.
+called on any expression whose compile-time type implements `java.io.Closeable`.
 
 The Units Checker supports new qualifiers (thanks to Rene Kraneis):
  * `@Volume`, `@m3`, `@mm3`, `@km3`