+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+Version 3.X.X (? ?, 2021)
+----------------------------
+
+**User-visible changes:**
+
+**Implementation details:**
+
+**Closed issues:**
+
 Version 3.14.0 (June 1, 2021)
 ----------------------------