+++ docs/CHANGELOG.md
@@ -5,11 +5,12 @@ Version 3.X.X (? ?, 2021)
 
 **Implementation details:**
 
-Renamed AnnotatedDeclaredType#wasRaw => #isUnderlyingTypeRaw and
-AnnotatedDeclaredType#setWasRaw => setIsUnderlyingTypeRaw. The old method
-remain but are deprecated.
+Method renamings (the old methods remain but are deprecated):
+ * `AnnotatedDeclaredType#wasRaw` => `isUnderlyingTypeRaw`
+ * `AnnotatedDeclaredType#setWasRaw` => `setIsUnderlyingTypeRaw`
 
 **Closed issues:**
+
 
 Version 3.14.0 (June 1, 2021)
 ----------------------------